Update MUL alias and add NEG alias in common_aliases.rs
Updated the MUL(Multiply) function alias to match the correct semantic of MUL and not MADD(Multiply-Add). Altered the documentation to reflect proper operational description, and changed params to fulfil the correct operational pseudocode for this instruction.

Additionally, Added NEG(Negate) function alias, a derivative of SUB (shifted register) method. Adjusted the documentation to state the correct functioning and alias relation with SUB method. Also implemented test for the newly added alias.

This update adheres to the accurate ARM documentation and ensures the proper functionality of the MUL and NEG instructions.