Add support for TBZ and TBNZ instructions
Added the new Imm14 and Offset16 variable types and implemented support for Test bit and Branch if Zero (TBZ) and Test bit and Branch if Nonzero (TBNZ) instructions in the assembly language. These bits-and-branch instructions compare a bit in a register to zero or not zero, and conditionally branch to a different part of the program, providing an important feature for control flow operations.